requirements.txt
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.6.0
SQLAlchemy==2.0.48
python-dotenv==1.1.1
ultralytics==8.4.33
opencv-python-headless==4.13.0.88
numpy==2.2.5
gunicorn==22.0.0
